feat: fix bugs and change name from asset to system token
